check if witness exists before deletion & rln full test
Former-commit-id: 5d56c23c79c1f4e6e4e6308f0f03fed605b9cdb7 [formerly eb1f3ab0cc06137f3adab8a6f230140652e164e6]
Former-commit-id: 4cdfffd291378f377208371af525384be33592c0

+++ packages/protocols/src/zk-protocol.ts
@@ -38,7 +38,8 @@ export class ZkProtocol {
   async genProof(grothInput: any, wasmFilePath: string, finalZkeyPath: string): Promise<FullProof> {
     await this.buildWnts(grothInput, wasmFilePath, 'witness.wtns');
     const { proof, publicSignals } = await groth16.prove(finalZkeyPath, 'witness.wtns', null);
-    fs.unlinkSync('witness.wtns');
+    const exists = fs.existsSync('witness.wtns');
+    if(exists) fs.unlinkSync('witness.wtns');
     return { proof, publicSignals };
   }
